requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.2.0
requests==2.32.3
whitenoise==6.7.0
gunicorn==22.0.0
django-cors-headers==4.4.0
